+++ src/babashka/impl/nrepl_server.clj
@@ -1,14 +1,21 @@
 (ns babashka.impl.nrepl-server
   {:no-doc true}
-  (:require [babashka.impl.common :as common]
-            [babashka.nrepl.server :as server]
-            [sci.core :as sci]))
+  (:require
+   [babashka.impl.clojure.core]
+   [babashka.impl.common :as common]
+   [babashka.nrepl.server :as server]
+   [sci.core :as sci]))
 
 (defn start-server!
   ([]
-   (server/start-server! @common/ctx))
+   (start-server! nil))
   ([opts]
-   (server/start-server! @common/ctx opts)))
+   (let [dev? (= "true" (System/getenv "BABASHKA_DEV"))
+         opts (merge {:debug dev?
+                      :describe {"versions" {"babashka" common/version}}
+                      :thread-bind [babashka.impl.clojure.core/warn-on-reflection]}
+                     opts)]
+     (server/start-server! @common/ctx opts))))
 
 (def nrepl-server-namespace
   (let [ns-sci (sci/create-ns 'babashka.nrepl.server)]