+++ CHANGELOG.md
@@ -12,6 +12,7 @@ A preview of the next release can be installed from
 - [#1369](https://github.com/babashka/babashka/issues/1369): provide `.sha256` files for every released asset
 - [#1397](https://github.com/babashka/babashka/issues/1397): Add `clojure.lang.Namespace` as alias for `sci.lang.Namespace`, such that `(instance? clojure.lang.Namespace *ns*)` returns `true` in bb
 - [#1384](https://github.com/babashka/babashka/issues/1384): allow `.indexOf` on `LazySeq`
+- [#1330](https://github.com/babashka/babashka/issues/1330): allow `(set! *warn-on-reflection*)` in programmatic nREPL
 
 ## 0.10.163 (2022-09-24)